chore: split Device into a separate module
